backend/requirements.txt
fastapi
uvicorn
python-dotenv
loguru
pytest
pydantic
openai
pinecone-client